+++ src/euphorie/client/tests/test_functional_risk.py
@@ -15,6 +15,11 @@ class RiskTests(EuphorieFunctionalTestCase):
         browser = self.get_browser()
         browser.open(survey.absolute_url())
         registerUserInClient(browser)
+        # We need to manually open the portlet view as the test browser
+        # does not handle JavaScript.
+        browser.open(
+            self.portal.client["nl"].absolute_url() + "/@@portlet-available-tools"
+        )
         # Create a new survey session
         browser.getControl(name="survey").value = ["ict/software-development"]
         browser.getForm(action="new-session").submit()
@@ -52,6 +57,11 @@ class RiskTests(EuphorieFunctionalTestCase):
         ].absolute_url()  # noqa: E501
         browser.open(survey_url)
         registerUserInClient(browser)
+        # We need to manually open the portlet view as the test browser
+        # does not handle JavaScript.
+        browser.open(
+            self.portal.client["nl"].absolute_url() + "/@@portlet-available-tools"
+        )
         # Create a new survey session
         browser.getControl(name="survey").value = ["ict/software-development"]
         browser.getForm(action="new-session").submit()
@@ -95,6 +105,11 @@ class RiskTests(EuphorieFunctionalTestCase):
         survey_url = self.portal.client.nl["ict"]["software-development"].absolute_url()
         browser.open(survey_url)
         registerUserInClient(browser)
+        # We need to manually open the portlet view as the test browser
+        # does not handle JavaScript.
+        browser.open(
+            self.portal.client["nl"].absolute_url() + "/@@portlet-available-tools"
+        )
         # Create a new survey session
         browser.getControl(name="survey").value = ["ict/software-development"]
         browser.getForm(action="new-session").submit()
@@ -133,6 +148,11 @@ class RiskTests(EuphorieFunctionalTestCase):
         ].absolute_url()  # noqa: E501
         browser.open(survey_url)
         registerUserInClient(browser)
+        # We need to manually open the portlet view as the test browser
+        # does not handle JavaScript.
+        browser.open(
+            self.portal.client["nl"].absolute_url() + "/@@portlet-available-tools"
+        )
         # Create a new survey session
         browser.getControl(name="survey").value = ["ict/software-development"]
         browser.getForm(action="new-session").submit()